switch from carthage to swift package manager
