--- Cartfile
@@ -1,1 +1,0 @@
-github "antitypical/Result" ~> 3.0
