--- Cartfile.private
@@ -1,2 +1,0 @@
-github "Quick/Nimble" "v5.1.1"
-github "ReactiveX/RxSwift" "3.0.1"
