--- Cartfile.resolved
@@ -1,3 +1,0 @@
-github "Quick/Nimble" "v5.1.1"
-github "antitypical/Result" "3.0.0"
-github "ReactiveX/RxSwift" "3.0.1"
