--- FaunaDB.xcworkspace/contents.xcworkspacedata
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Workspace
-   version = "1.0">
-   <FileRef
-      location = "container:FaunaDB.xcodeproj">
-   </FileRef>
-   <FileRef
-      location = "group:Example.xcodeproj">
-   </FileRef>
-</Workspace>
