--- FaunaDB/FaunaDB.h
@@ -1,18 +1,0 @@
-//
-//  FaunaDB.h
-//  FaunaDB
-//
-//  Copyright © 2016 Fauna, Inc. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for FaunaDB.
-FOUNDATION_EXPORT double FaunaDBVersionNumber;
-
-//! Project version string for FaunaDB.
-FOUNDATION_EXPORT const unsigned char FaunaDBVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <FaunaDB/PublicHeader.h>
-
-
