+++ Sources/FaunaDB/Errors.swift
