+++ Sources/FaunaDB/Expr.swift
@@ -1,5 +1,52 @@
 import Foundation
 
+public protocol Expr {}
+
+precedencegroup ExprTuplePrecedence {
+    assignment: true
+    associativity: left
+    lowerThan: CastingPrecedence
+}
+
+infix operator => : ExprTuplePrecedence
+public func => (key: String, value: Expr?) -> (String, Expr?) {
+    return (key, value)
+}
+
+protocol Fn: Expr, AsJson, CustomStringConvertible {
+    typealias Call = [String: Expr]
+    var call: Call { get }
+}
+
+internal extension Fn {
+    func escape() -> JsonType {
+        return .object(call.mapValuesT(JSON.escape))
+    }
+}
+
+extension Fn {
+    public var description: String {
+        return call.description
+    }
+}
+
+internal func fn(_ pairs: (String, Expr?)...) -> Fn.Call {
+    return Dictionary(pairs:
+        pairs.flatMap { (key, value) in
+            guard let value = value else { return nil }
+            return (key, value)
+        }
+    )
+}
+
+internal func varargs(_ args: [Expr]) -> Expr {
+    if args.count == 1 {
+        return args.first!
+    }
+
+    return Arr(wrap: args)
+}
+
 public struct Obj: Expr, AsJson, CustomStringConvertible {
 
     private let wrapped: [String: Expr?]
@@ -42,20 +89,6 @@ public struct Arr: Expr, AsJson, CustomStringConvertible {
 
     func escape() -> JsonType {
         return .array(wrapped.map(JSON.escape))
-    }
-
-}
-
-public struct Ref: Fn {
-
-    var call: Fn.Call
-
-    public init(_ id: String) {
-        self.call = fn("@ref" => id)
-    }
-
-    public init(class: Expr, id: Expr) {
-        self.call = fn("ref" => `class`, "id" => id)
     }
 
 }
