+++ Sources/FaunaDB/Path.swift
@@ -1,5 +1,3 @@
-import Foundation
-
 internal struct Path {
 
     static let root = Path([])
