+++ Sources/FaunaDB/QueryResult.swift
