+++ Tests/FaunaDBTests/ErrorsTests.swift
