--- Tests/TestConfig.plist
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>FAUNA_ROOT_KEY</key>
-	<string>Add your cloud secret here</string>
-</dict>
-</plist>
